+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Tue Apr 09 09:03:59 CEST 2019
-gradle.version=4.6
+#Mon May 06 12:32:53 CEST 2019
+gradle.version=5.1.1
