+++ .idea/gradle.xml
@@ -3,12 +3,15 @@
   <component name="GradleSettings">
     <option name="linkedExternalProjectsSettings">
       <GradleProjectSettings>
+        <compositeConfiguration>
+          <compositeBuild compositeDefinitionSource="SCRIPT" />
+        </compositeConfiguration>
         <option name="distributionType" value="DEFAULT_WRAPPED" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
         <option name="modules">
           <set>
-            <option value="$PROJECT_DIR$/../../prog-unbook/SystemetAppBasic" />
-            <option value="$PROJECT_DIR$/../../prog-unbook/SystemetAppBasic/app" />
+            <option value="$USER_HOME$/prog-unbook/SystemetAppBasic" />
+            <option value="$USER_HOME$/prog-unbook/SystemetAppBasic/app" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
