+++ app/src/main/res/layout/activity_main.xml
@@ -1,16 +1,115 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".MainActivity">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
 
-  <ListView
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:id="@+id/product_list">
-  </ListView>
+  <TextView
+      android:id="@+id/textView"
+      android:layout_width="344dp"
+      android:layout_height="151dp"
+      android:text="This app contains information about alcoholic beverages for visit this app you must be over 20 years old"
+      android:textAlignment="center"
+      android:textSize="24sp"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintHorizontal_bias="0.597"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.443" />
 
+  <Button
+      android:id="@+id/button2"
+      android:layout_width="160dp"
+      android:layout_height="46dp"
+      android:background="@color/colorPrimary"
+      android:text="Yes, I am over 20 or older"
+      android:textAlignment="center"
+      android:textSize="14sp"
+      android:textStyle="bold"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintHorizontal_bias="0.103"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.645" />
+
+  <Button
+      android:id="@+id/button1"
+      android:layout_width="160dp"
+      android:layout_height="46dp"
+      android:background="@color/colorPrimary"
+      android:text="I am not old enough"
+      android:textAlignment="center"
+      android:textSize="14sp"
+      android:textStyle="bold"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintHorizontal_bias="0.828"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.645" />
+
+  <ImageView
+      android:id="@+id/imageView"
+      android:layout_width="346dp"
+      android:layout_height="112dp"
+      android:layout_marginStart="25dp"
+      android:layout_marginEnd="25dp"
+      android:layout_marginBottom="16dp"
+      app:layout_constraintBottom_toTopOf="@+id/textView"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:srcCompat="@drawable/liquor" />
+
+  <TextView
+      android:id="@+id/textView3"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:fontFamily="cursive"
+      android:text="Do you like this app?"
+      android:textSize="24sp"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintHorizontal_bias="0.497"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.724" />
+
+  <RatingBar
+      android:id="@+id/RatingBar"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:numStars="5"
+      android:rating="0"
+      android:stepSize="1"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintHorizontal_bias="0.497"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.799" />
+
+
+  <Button
+      android:id="@+id/button3"
+      android:layout_width="30dp"
+      android:layout_height="21dp"
+      android:layout_marginStart="162dp"
+      android:layout_marginTop="596dp"
+      android:layout_marginEnd="161dp"
+      android:layout_marginBottom="87dp"
+      android:text="Submit"
+      android:textAlignment="center"
+      android:textColor="@color/colorAccent"
+      android:textSize="12sp"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent" />
+
+  />
 
 </android.support.constraint.ConstraintLayout>
